#5 renamed scikits.odes classes and added installation of sundials + scikits.odes in travis
