+++ tests/test_solvers/test_scikits_solvers.py
@@ -1,18 +1,23 @@
 #
-# Tests for the Odes Solver class
+# Tests for the Scikits Solver class
 #
 from __future__ import absolute_import, division
 from __future__ import print_function, unicode_literals
 import pybamm
+from pybamm.solvers.scikits_ode_solver import scikits_odes_spec
+
 
 import unittest
 import numpy as np
 
 
-class TestOdesSolver(unittest.TestCase):
+class TestScikitsSolver(unittest.TestCase):
+
+    @unittest.skipIf(scikits_odes_spec is None,
+                     "scikits.odes not installed")
     def test_ode_integrate(self):
         # Constant
-        solver = pybamm.OdesOdeSolver(tol=1e-8)
+        solver = pybamm.ScikitsOdeSolver(tol=1e-8)
 
         def constant_growth(t, y):
             return 0.5 * np.ones_like(y)
@@ -24,7 +29,7 @@ class TestOdesSolver(unittest.TestCase):
         np.testing.assert_allclose(0.5 * t_sol, y_sol[0])
 
         # Exponential decay
-        solver = pybamm.OdesOdeSolver(tol=1e-8)
+        solver = pybamm.ScikitsOdeSolver(tol=1e-8)
 
         def exponential_decay(t, y):
             return -0.1 * y
@@ -34,9 +39,11 @@ class TestOdesSolver(unittest.TestCase):
         t_sol, y_sol = solver.integrate(exponential_decay, y0, t_eval)
         np.testing.assert_allclose(y_sol[0], np.exp(-0.1 * t_sol))
 
+    @unittest.skipIf(scikits_odes_spec is None,
+                     "scikits.odes not installed")
     def test_dae_integrate(self):
         # Constant
-        solver = pybamm.OdesDaeSolver(tol=1e-8)
+        solver = pybamm.ScikitsDaeSolver(tol=1e-8)
 
         def constant_growth_dae(t, y, ydot):
             return [
@@ -53,7 +60,7 @@ class TestOdesSolver(unittest.TestCase):
         np.testing.assert_allclose(1.0 * t_sol, y_sol[1])
 
         # Exponential decay
-        solver = pybamm.OdesDaeSolver(tol=1e-8)
+        solver = pybamm.ScikitsDaeSolver(tol=1e-8)
 
         def exponential_decay_dae(t, y, ydot):
             return [
@@ -69,6 +76,8 @@ class TestOdesSolver(unittest.TestCase):
         np.testing.assert_allclose(y_sol[0], np.exp(-0.1 * t_sol))
         np.testing.assert_allclose(y_sol[1], 2*np.exp(-0.1 * t_sol))
 
+    @unittest.skipIf(scikits_odes_spec is None,
+                     "scikits.odes not installed")
     def test_model_solver_ode(self):
         # Create model
         model = pybamm.BaseModel()
@@ -84,12 +93,14 @@ class TestOdesSolver(unittest.TestCase):
         disc.process_model(model)
 
         # Solve
-        solver = pybamm.OdesOdeSolver(tol=1e-9)
+        solver = pybamm.ScikitsOdeSolver(tol=1e-9)
         t_eval = mesh["time"]
         solver.solve(model, t_eval)
         np.testing.assert_array_equal(solver.t, t_eval)
         np.testing.assert_allclose(solver.y[0], np.exp(0.1 * solver.t))
 
+    @unittest.skipIf(scikits_odes_spec is None,
+                     "scikits.odes not installed")
     def test_model_solver_dae(self):
         # Create model
         model = pybamm.BaseModel()
@@ -110,7 +121,7 @@ class TestOdesSolver(unittest.TestCase):
         disc.process_model(model)
 
         # Solve
-        solver = pybamm.OdesDaeSolver(tol=1e-8)
+        solver = pybamm.ScikitsDaeSolver(tol=1e-8)
         t_eval = mesh["time"]
         solver.solve(model, t_eval)
         np.testing.assert_array_equal(solver.t, t_eval)
